broke up project into scenes, add start button place holder
