+++ client/index.html
@@ -2,6 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8">
+    <title>WarTank</title>
   </head>
   <body>
   </body>
