+++ client/webpack/base.js
@@ -37,7 +37,7 @@ module.exports = {
       WEBGL_RENDERER: JSON.stringify(true)
     }),
     new HtmlWebpackPlugin({
-      template: "./index.html"
+      template: "./index.html",
     })
   ]
 };
